testscripts/requirements.txt
boto3==1.4.0
botocore==1.4.92
click==6.6
configparser==3.5.0
cycler==0.10.0
docutils==0.13.1
Flask==0.12
itsdangerous==0.24
Jinja2==2.8.1
jmespath==0.9.0
kafka-python==1.3.2
MarkupSafe==0.23
matplotlib==1.5.3
mysqlclient==1.3.9
numpy==1.11.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.11.1
s3transfer==0.1.10
six==1.10.0
uWSGI==2.0.14
Werkzeug==0.11.15
